refactor: move Icon component and animation-utils to shared locations, remove @kiku/core dependency
