+++ tanaka/extension/src/components/icon/icon.scss
