+++ tanaka/extension/src/components/icon/index.ts
